# gonna-pay-backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateGroupRequest:
    properties:
      category:
        enum:
        - Dinner
        - Lunch
        - Entertainment
        - Travel
        - Others
        type: string
      memberIds:
        items:
          type: string
        type: array
      name:
        type: string
    required:
    - category
    - name
    type: object
...
  request.UpdateGroupRequest:
    properties:
      category:
        enum:
        - Dinner
        - Lunch
        - Entertainment
        - Travel
        - Others
        type: string
      name:
        type: string
    type: object
...
  response.GroupDetailResponse:
    properties:
      category:
        type: string
      createdAt:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/response.MemberResponse'
        type: array
      name:
        type: string
      updatedAt:
        type: string
    type: object
  response.GroupResponse:
    properties:
      category:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  response.MemberResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  response.PaginatedGroupResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.GroupResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  rest_errors.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_errors.RestErrors:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_errors.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:3333
info:
  contact: {}
  description: API de gestão financeira e divisão de custos
  title: Gonna Pay API
  version: "1.0"
paths:
...
  /group:
    post:
      consumes:
      - application/json
      description: Cria um novo grupo. É possível já adicionar membros (contatos)
        pelo campo memberIds
      parameters:
      - description: Dados do grupo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Criar grupo
      tags:
      - groups
  /group/{id}:
    delete:
      description: Remove um grupo. Apenas o dono pode deletar
      parameters:
      - description: ID do grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Deletar grupo
      tags:
      - groups
    get:
      description: Retorna um grupo com a lista de membros. Apenas o dono pode acessar
      parameters:
      - description: ID do grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GroupDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Grupo não pertence ao usuário
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Buscar grupo por ID
      tags:
      - groups
    patch:
      consumes:
      - application/json
      description: Atualiza nome e/ou categoria do grupo. Apenas o dono pode editar
      parameters:
      - description: ID do grupo
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Atualizar grupo
      tags:
      - groups
  /group/{id}/member/{contactId}:
    delete:
      description: Remove um contato do grupo. Apenas o dono do grupo pode remover
      parameters:
      - description: ID do grupo
        in: path
        name: id
        required: true
        type: string
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Remover membro
      tags:
      - groups
    post:
      description: Adiciona um contato como membro do grupo. Apenas o dono do grupo
        pode adicionar
      parameters:
      - description: ID do grupo
        in: path
        name: id
        required: true
        type: string
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "409":
          description: Membro já está no grupo
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Adicionar membro
      tags:
      - groups
  /groups:
    get:
      description: Retorna todos os grupos do usuário autenticado com paginação
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 20, máximo: 100)'
        in: query
        name: limit
        type: integer
      - description: Filtrar por categoria
        enum:
        - Dinner
        - Lunch
        - Entertainment
        - Travel
        - Others
        in: query
        name: category
        type: string
      - description: Buscar por nome
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedGroupResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Listar grupos
      tags:
      - groups
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"